Split python backend
Split into separate server and crawler implementation modules
